requirements.txt
torch==2.1.0
torchvision==0.17.1
numpy==1.25.2
opencv-python==4.7.0.72
Pillow==10.0.0